(chores) camel-paho-mqtt5: fix integration/unit test mismatch
- should ensure ITs are properly skipped on s390x
- ensure the ITs run with failsafe
- ensure consistency with other components
