+++ components/camel-paho-mqtt5/src/test/java/org/apache/camel/component/paho/mqtt5/integration/PahoMqtt5OverrideTopicMqtt5IT.java
@@ -14,24 +14,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.paho.mqtt5;
+package org.apache.camel.component.paho.mqtt5.integration;
 
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
+import org.apache.camel.component.paho.mqtt5.PahoMqtt5Component;
+import org.apache.camel.component.paho.mqtt5.PahoMqtt5Constants;
 import org.junit.jupiter.api.Test;
 
-public class PahoMqtt5ToDTest extends PahoMqtt5TestSupport {
-
-    @Test
-    public void testToD() throws Exception {
-        getMockEndpoint("mock:bar").expectedBodiesReceived("Hello bar");
-        getMockEndpoint("mock:beer").expectedBodiesReceived("Hello beer");
-
-        template.sendBodyAndHeader("direct:start", "Hello bar", "where", "bar");
-        template.sendBodyAndHeader("direct:start", "Hello beer", "where", "beer");
-
-        MockEndpoint.assertIsSatisfied(context);
-    }
+public class PahoMqtt5OverrideTopicMqtt5IT extends PahoMqtt5ITSupport {
 
     @Override
     protected RouteBuilder createRouteBuilder() {
@@ -41,13 +32,25 @@ public class PahoMqtt5ToDTest extends PahoMqtt5TestSupport {
                 PahoMqtt5Component paho = context.getComponent("paho-mqtt5", PahoMqtt5Component.class);
                 paho.getConfiguration().setBrokerUrl("tcp://localhost:" + mqttPort);
 
-                // route message dynamic using toD
-                from("direct:start").toD("paho-mqtt5:${header.where}");
+                from("direct:test").to("paho-mqtt5:queue").log("Message sent");
 
-                from("paho-mqtt5:bar").to("mock:bar");
-                from("paho-mqtt5:beer").to("mock:beer");
+                from("paho-mqtt5:myoverride").log("Message received").to("mock:test");
             }
         };
+    }
+
+    // Tests
+
+    @Test
+    public void shouldOverride() throws InterruptedException {
+        // Given
+        getMockEndpoint("mock:test").expectedMessageCount(1);
+
+        // When
+        template.sendBodyAndHeader("direct:test", "Hello World", PahoMqtt5Constants.CAMEL_PAHO_OVERRIDE_TOPIC, "myoverride");
+
+        // Then
+        MockEndpoint.assertIsSatisfied(context);
     }
 
 }
